requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
sklearn.model_selection
sklearn.metrics
sklearn.linear_model
sklearn.ensemble
sklearn.tree
sklearn.svm
sklearn.naive_bayes
sklearn.neighbors
pandas
plotly
streamlit
yfinance
pandas-datareader
prophet
